+++ tools/tsp-client/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release
 
+## 2024-12-10 - 0.15.0
+
+- Support specifying an entrypoint file in tsp-location.yaml.
+- Bug fix: give preference to client.tsp over main.tsp in the entrypoint file search.
+
 ## 2024-12-03 - 0.14.2
 
 - Bumped `@autorest/openapi-to-typespec` version to `0.10.4`.